added const&'s and such; added size_t's

+++ mode.h
@@ -7,7 +7,7 @@
 
 class Mode {
 public:
-    virtual std::vector<unsigned int> getOrder(std::vector<std::shared_ptr<Playable>> list) = 0;
+    virtual std::vector<size_t> getOrder(std::vector<std::shared_ptr<Playable>> list) = 0;
     virtual ~Mode() = default;
 };
